feat: added all building groups query
